requirements.txt
PyYAML
pysnmplib